+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.14] - 2025-11-03
+### Fixed
+- **Module Management:**
+  - Fixed bug where `Manage Modules` button did not open the enhanced module management window when other core language was set. 
+
 ## [0.5.13] - 2025-11-02
 ### Fixed
 - **Localization:**